refactor: split _svar.py into _svar/ package
Pure code motion along lifecycle seams (_regions, _convert, _io, _kernels,
_annotate mixin, _core) with a re-export shim. No public-API or behavior
change; existing suite is the characterization gate.
