+++ python/genoray/_svar/__init__.py
@@ -1,0 +1,4 @@
+from ._convert import dense2sparse
+from ._core import SparseVar, SparseVarMetadata
+
+__all__ = ["SparseVar", "SparseVarMetadata", "dense2sparse"]
