+++ tests/test_svar_from_subset.py
@@ -10,12 +10,9 @@ import subprocess
 import pytest
 
 from genoray import PGEN, VCF, SparseVar
-from genoray._svar import (
-    V_IDX_TYPE,
-    _build_working_index,
-    _resolve_kept_rows,
-    _subset_var_idxs_and_recompute_af,
-)
+from genoray._svar._io import _build_working_index, _subset_var_idxs_and_recompute_af
+from genoray._svar._regions import _resolve_kept_rows
+from genoray._types import V_IDX_TYPE
 from genoray._utils import ContigNormalizer
